Fix name miscompare
Signed-off-by: Naresh Bhat<naresh.bhat@linaro.org>

On big-endian machines, a test case looking for the methods _L1D and _E1D
in the same scope would fail (see tests/misc/badcode.asl:184).  The names
to be compared were being treated as 32-bit ints, and not strings.  Hence,
the characters were re-ordered incorrectly, mismatching the assumptions
made in the remainder of the function.